requirements.txt
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
streamlit==1.25.0
plotly==5.15.0
scikit-learn==1.3.0